// frontend/app/@types/schemas.d.ts
export interface ApiArticleArticle extends CollectionTypeSchema {
  info: {
    singularName: "article";
    pluralName: "articles";
    displayName: "Article";
    description: "";
  };
  options: {
    draftAndPublish: true;
  };
  pluginOptions: {
    i18n: {
      localized: true;
    };
  };
  attributes: {
    title: StringAttribute &
      RequiredAttribute &
      UniqueAttribute &
      SetPluginOptions<{
        i18n: {
          localized: true;
        };
      }>;
    text: RichTextAttribute &
      RequiredAttribute &
      SetPluginOptions<{
        i18n: {
          localized: true;
        };
      }>;
    summary: TextAttribute &
      RequiredAttribute &
      SetPluginOptions<{
        i18n: {
          localized: true;
        };
      }>;
    slug: StringAttribute &
      RequiredAttribute &
      UniqueAttribute &
      SetPluginOptions<{
        i18n: {
          localized: true;
        };
      }>;
    image: MediaAttribute &
      SetPluginOptions<{
        i18n: {
          localized: true;
        };
      }>;
    createdAt: DateTimeAttribute;
    updatedAt: DateTimeAttribute;
    publishedAt: DateTimeAttribute;
    createdBy: RelationAttribute<
      "api::article.article",
      "oneToOne",
      "admin::user"
    > &
      PrivateAttribute;
    updatedBy: RelationAttribute<
      "api::article.article",
      "oneToOne",
      "admin::user"
    > &
      PrivateAttribute;
    localizations: RelationAttribute<
      "api::article.article",
      "oneToMany",
      "api::article.article"
    >;
    locale: StringAttribute;
  };
}

export interface ApiFooterFooter extends SingleTypeSchema {
  info: {
    singularName: "footer";
    pluralName: "footers";
    displayName: "Footer";
  };
  options: {
    draftAndPublish: true;
  };
  pluginOptions: {
    i18n: {
      localized: true;
    };
  };
  attributes: {
    copyright: StringAttribute &
      SetPluginOptions<{
        i18n: {
          localized: true;
        };
      }>;
    contact: StringAttribute &
      SetPluginOptions<{
        i18n: {
          localized: true;
        };
      }>;
    createdAt: DateTimeAttribute;
    updatedAt: DateTimeAttribute;
    publishedAt: DateTimeAttribute;
    createdBy: RelationAttribute<
      "api::footer.footer",
      "oneToOne",
      "admin::user"
    > &
      PrivateAttribute;
    updatedBy: RelationAttribute<
      "api::footer.footer",
      "oneToOne",
      "admin::user"
    > &
      PrivateAttribute;
    localizations: RelationAttribute<
      "api::footer.footer",
      "oneToMany",
      "api::footer.footer"
    >;
    locale: StringAttribute;
  };
}

export interface ApiHeaderHeader extends SingleTypeSchema {
  info: {
    singularName: "header";
    pluralName: "headers";
    displayName: "Header";
    description: "";
  };
  options: {
    draftAndPublish: true;
  };
  pluginOptions: {
    i18n: {
      localized: true;
    };
  };
  attributes: {
    siteName: StringAttribute &
      SetPluginOptions<{
        i18n: {
          localized: true;
        };
      }>;
    logo: MediaAttribute &
      SetPluginOptions<{
        i18n: {
          localized: true;
        };
      }>;
    subtitle: StringAttribute &
      SetPluginOptions<{
        i18n: {
          localized: true;
        };
      }>;
    createdAt: DateTimeAttribute;
    updatedAt: DateTimeAttribute;
    publishedAt: DateTimeAttribute;
    createdBy: RelationAttribute<
      "api::header.header",
      "oneToOne",
      "admin::user"
    > &
      PrivateAttribute;
    updatedBy: RelationAttribute<
      "api::header.header",
      "oneToOne",
      "admin::user"
    > &
      PrivateAttribute;
    localizations: RelationAttribute<
      "api::header.header",
      "oneToMany",
      "api::header.header"
    >;
    locale: StringAttribute;
  };
}